+++ database/migrations/2024_06_10_190316_create_to_be_read_list.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('to_be_read_list', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id');
-            $table->integer('book_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('book_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->foreign('book_id')->references('id')->on('books')->onDelete('cascade');
             $table->timestamps();
